requirements.txt
Pillow
django-ckeditor